+++ app/src/main/AndroidManifest.xml
@@ -7,6 +7,9 @@
     <!-- This is for API below 21 -->
     <uses-permission android:name="android.permission.CAMERA"/>
     <uses-permission android:name="android.permission.FOREGROUND_SERVICE"/>
+    <uses-permission
+        android:name="android.permission.WAKE_LOCK"
+        tools:node="replace" />
 
     <application
         android:allowBackup="true"
@@ -28,6 +31,7 @@
         <activity
             android:name=".NewUserActivity"
             android:screenOrientation="portrait"
+            android:exported="false"
             android:label="@string/app_name"
             android:windowSoftInputMode="adjustPan"
             android:theme="@style/AppTheme.NoActionBar">
@@ -35,21 +39,27 @@
         <activity
             android:name=".MainActivity"
             android:screenOrientation="portrait"
+            android:exported="false"
             android:label="@string/app_name"
             android:theme="@style/AppTheme.NoActionBar">
         </activity>
         <activity
             android:name=".SettingsActivity"
             android:screenOrientation="portrait"
+            android:exported="false"
             android:label="@string/app_name"
             android:theme="@style/AppTheme.NoActionBar">
         </activity>
         <activity
             android:name="com.journeyapps.barcodescanner.CaptureActivity"
             android:screenOrientation="portrait"
+            android:exported="false"
             tools:replace="android:screenOrientation"
             android:stateNotNeeded="true"/>
-        <service android:name=".service.YourService"/>
+        <service
+            android:name=".wallet.WalletService"
+            android:exported="false"
+            android:stopWithTask="true"/>
     </application>
 
 </manifest>
