+++ app/src/main/java/xyz/pokkst/pokket/cash/ui/main/fragment/setting/SettingsPhraseFragment.kt
@@ -7,7 +7,7 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import kotlinx.android.synthetic.main.fragment_settings_phrase.view.*
 import xyz.pokkst.pokket.cash.R
-import xyz.pokkst.pokket.cash.wallet.WalletManager
+import xyz.pokkst.pokket.cash.wallet.WalletService
 
 
 /**
@@ -20,7 +20,7 @@ class SettingsPhraseFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         val root = inflater.inflate(R.layout.fragment_settings_phrase, container, false)
-        root.the_phrase.text = WalletManager.wallet?.keyChainSeed?.mnemonicString
+        root.the_phrase.text = WalletService.wallet?.keyChainSeed?.mnemonicString
         return root
     }
 }
